Feat: did a tidy clean up of code and making it modular
