+++ src/JS_Modules/utils.js
@@ -1,5 +1,7 @@
 import Likes from './likes.js';
 import Comments from './comment.js';
+import { Comment } from './components.js';
+
 /**
  * @class Utils - the utility class to hold all utiltity functionality
  */
@@ -31,11 +33,8 @@ export default class Utils {
    * @returns {Number} the length of the Array
    * @memberof Utils
    */
-  static getCount(data) {
-    if (Array.isArray(data)) {
-      return data.length;
-    }
-    return 0;
+  static getCount(data = []) {
+    return data.length;
   }
 
   /**
@@ -71,8 +70,40 @@ export default class Utils {
     setTimeout(() => elem.classList.remove(type), 800);
   }
 
-  static async displayComments(id) {
+  
+  static async getComments(id) {
     const ans = await Comments.getCommentOne(id);
-    return ans.map((item) => item);
+    return ans.error ? [] : ans.map((item) => item);
+  }
+
+  static async display(args) {
+    const data = await Utils.getComments(args);
+    const space = document.querySelector('.comment-display');
+    const title = document.querySelector('.comment-section h3');
+    const count = Utils.getCount(data);
+    title.innerHTML = `Comments(${count})`;
+    space.innerHTML = data.map((item) => Comment(item)).join('');
+  }
+
+  static handleClose(modal) {
+    const span = document.querySelector('.close');
+    span.addEventListener('click', () => {
+      modal.style.display = 'none';
+      modal.innerHTML = '';
+    });
+  }
+
+  static handleForm(args) {
+    const forms = document.getElementById('form');
+    forms.addEventListener('submit', async (e) => {
+      e.preventDefault();
+      const name = document.getElementById('name');
+      const comment = document.getElementById('comment');
+      if ((name.value !== '') && (comment.value !== '')) {
+        await Comments.postComment(args, name.value, comment.value);
+        Utils.display(args);
+        forms.reset();
+      }
+    });
   }
 }
