+++ src/__mocks__/fileMock.js
@@ -1,0 +1,1 @@
+export default '';
