+++ webpack.config.js
@@ -5,7 +5,6 @@ module.exports = {
   mode: 'development',
   entry: {
     index: './src/index.js',
-    helper: './src/helper.js',
   },
   devtool: 'inline-source-map',
   devServer: {
